+++ tests/test_alphabet.py
@@ -6,14 +6,14 @@ def test_simple():
     sqids = Sqids(alphabet="0123456789abcdef")
 
     numbers = [1, 2, 3]
-    id = "4d9fd2"
+    id = "489158"
 
     assert sqids.encode(numbers) == id
     assert sqids.decode(id) == numbers
 
 
 def test_short_alphabet():
-    sqids = Sqids(alphabet="abcde")
+    sqids = Sqids(alphabet="abc")
 
     numbers = [1, 2, 3]
     assert sqids.decode(sqids.encode(numbers)) == numbers
@@ -28,6 +28,11 @@ def test_long_alphabet():
     assert sqids.decode(sqids.encode(numbers)) == numbers
 
 
+def test_multibyte_alphabet():
+    with pytest.raises(Exception):
+        Sqids(alphabet="ë1092")
+
+
 def test_repeating_alphabet_characters():
     with pytest.raises(Exception):
         Sqids(alphabet="aabcdefg")
@@ -35,4 +40,4 @@ def test_repeating_alphabet_characters():
 
 def test_too_short_alphabet():
     with pytest.raises(Exception):
-        Sqids(alphabet="abcd")
+        Sqids(alphabet="ab")