+++ tests/test_encoding.py
@@ -1,34 +1,35 @@
 import pytest
+import sys
 from sqids import Sqids
 
 
 def test_simple():
     sqids = Sqids()
     numbers = [1, 2, 3]
-    id_str = "8QRLaD"
+    id_str = "86Rf07"
     assert sqids.encode(numbers) == id_str
     assert sqids.decode(id_str) == numbers
 
 
 def test_different_inputs():
     sqids = Sqids()
-    numbers = [0, 0, 0, 1, 2, 3, 100, 1_000, 100_000, 1_000_000, sqids.max_value()]
+    numbers = [0, 0, 0, 1, 2, 3, 100, 1_000, 100_000, 1_000_000, sys.maxsize]
     assert sqids.decode(sqids.encode(numbers)) == numbers
 
 
 def test_incremental_numbers():
     sqids = Sqids()
     ids = {
-        "bV": [0],
-        "U9": [1],
-        "g8": [2],
-        "Ez": [3],
-        "V8": [4],
-        "ul": [5],
-        "O3": [6],
-        "AF": [7],
-        "ph": [8],
-        "n8": [9],
+        "bM": [0],
+        "Uk": [1],
+        "gb": [2],
+        "Ef": [3],
+        "Vq": [4],
+        "uw": [5],
+        "OI": [6],
+        "AX": [7],
+        "p6": [8],
+        "nJ": [9],
     }
     for id_str, numbers in ids.items():
         assert sqids.encode(numbers) == id_str
@@ -38,16 +39,16 @@ def test_incremental_numbers():
 def test_incremental_numbers_same_index_0():
     sqids = Sqids()
     ids = {
-        "SrIu": [0, 0],
-        "nZqE": [0, 1],
-        "tJyf": [0, 2],
-        "e86S": [0, 3],
-        "rtC7": [0, 4],
-        "sQ8R": [0, 5],
-        "uz2n": [0, 6],
-        "7Td9": [0, 7],
-        "3nWE": [0, 8],
-        "mIxM": [0, 9],
+        "SvIz": [0, 0],
+        "n3qa": [0, 1],
+        "tryF": [0, 2],
+        "eg6q": [0, 3],
+        "rSCF": [0, 4],
+        "sR8x": [0, 5],
+        "uY2M": [0, 6],
+        "74dI": [0, 7],
+        "30WX": [0, 8],
+        "moxr": [0, 9],
     }
     for id_str, numbers in ids.items():
         assert sqids.encode(numbers) == id_str
@@ -57,16 +58,16 @@ def test_incremental_numbers_same_index_0():
 def test_incremental_numbers_same_index_1():
     sqids = Sqids()
     ids = {
-        "SrIu": [0, 0],
-        "nbqh": [1, 0],
-        "t4yj": [2, 0],
-        "eQ6L": [3, 0],
-        "r4Cc": [4, 0],
-        "sL82": [5, 0],
-        "uo2f": [6, 0],
-        "7Zdq": [7, 0],
-        "36Wf": [8, 0],
-        "m4xT": [9, 0],
+        "SvIz": [0, 0],
+        "nWqP": [1, 0],
+        "tSyw": [2, 0],
+        "eX68": [3, 0],
+        "rxCY": [4, 0],
+        "sV8a": [5, 0],
+        "uf2K": [6, 0],
+        "7Cdk": [7, 0],
+        "3aWP": [8, 0],
+        "m2xn": [9, 0],
     }
     for id_str, numbers in ids.items():
         assert sqids.encode(numbers) == id_str
@@ -90,11 +91,6 @@ def test_decoding_empty_string():
     assert sqids.decode("") == []
 
 
-def test_decoding_bad_id_with_repeating_reserved_chars():
-    sqids = Sqids()
-    assert sqids.decode("fff") == []
-
-
 def test_decoding_invalid_character():
     sqids = Sqids()
     assert sqids.decode("*") == []
@@ -103,6 +99,6 @@ def test_decoding_invalid_character():
 def test_encode_out_of_range_numbers():
     sqids = Sqids()
     with pytest.raises(ValueError):
-        sqids.encode([sqids.min_value() - 1])
+        sqids.encode([-1])
     with pytest.raises(ValueError):
-        sqids.encode([sqids.max_value() + 1])
+        sqids.encode([sys.maxsize + 1])